+++ test/functional/abc_rpc_finalizetransaction.py
@@ -1,7 +1,7 @@
 # Copyright (c) 2025 The Bitcoin developers
 # Distributed under the MIT software license, see the accompanying
 # file COPYING or http://www.opensource.org/licenses/mit-license.php.
-"""Test the finalizetransaction RPC."""
+"""Test the finalizetransaction and removetransaction RPCs."""
 from test_framework.p2p import P2PInterface
 from test_framework.test_framework import BitcoinTestFramework
 from test_framework.util import assert_equal, assert_raises_rpc_error
@@ -53,11 +53,19 @@ class AvalancheFinalizeTransactionTest(BitcoinTestFramework):
         # No tx in the block template
         assert_gbt_transaction_ids([])
 
-        # Finalizing a non existing transactions returns an error
+        # Finalizing a non existing transaction returns an error
         assert_raises_rpc_error(
             -8,
             "The transaction is not in the mempool",
             node.finalizetransaction,
+            "00" * 32,
+        )
+
+        # Removing a non existing trnsaction returns an error
+        assert_raises_rpc_error(
+            -8,
+            "The transaction is not in the mempool",
+            node.removetransaction,
             "00" * 32,
         )
 
@@ -69,6 +77,12 @@ class AvalancheFinalizeTransactionTest(BitcoinTestFramework):
         assert_equal(node.finalizetransaction(txids[1]), [txids[1]])
         assert_equal(node.finalizetransaction(txids[2]), [txids[2]])
         assert_gbt_transaction_ids(txids[:3])
+
+        # Remove the first transaction, it should no longer be in the block
+        # template
+        assert_equal(node.removetransaction(txids[0]), [txids[0]])
+        assert txids[0] not in node.getrawmempool()
+        assert_gbt_transaction_ids(txids[1:3])
 
         chained_txs = wallet.send_self_transfer_chain(from_node=node, chain_length=5)
         assert all(tx["txid"] in node.getrawmempool() for tx in chained_txs)
@@ -82,7 +96,22 @@ class AvalancheFinalizeTransactionTest(BitcoinTestFramework):
         )
 
         # They are all included in the block template but no other
-        assert_gbt_transaction_ids(txids[:3] + expected_chained_txids)
+        assert_gbt_transaction_ids(txids[1:3] + expected_chained_txids)
+
+        # Remove the second transaction from the chain, check the descendants
+        # are removed as well and no longer in the block template
+        expected_removed_chained_txids = [tx["txid"] for tx in chained_txs[1:]]
+        assert_equal(
+            sorted(node.removetransaction(chained_txs[1]["txid"])),
+            sorted(expected_removed_chained_txids),
+        )
+        assert all(
+            txid not in node.getrawmempool() for txid in expected_removed_chained_txids
+        )
+        # Only the first tx from the chain remain in the mempool and in the
+        # block template
+        assert chained_txs[0]["txid"] in node.getrawmempool()
+        assert_gbt_transaction_ids(txids[1:3] + [chained_txs[0]["txid"]])
 
 
 if __name__ == "__main__":